Crud is complete on pages module
Alex and I were able to write working create, index, show, update, and
destroy for the pages model. First we went through and changed the
examples controller and model to represent the page. Next one by one
we went through the each script and then editing the route to accept the
new script. We went one by one, testing each along the way.

Next step we're finishing the blog model and then connecting them both
in order to test the code before involving the front end.