Fixed unauthorized show and index.
Took out the ability to index or show with out authorization. Did so by
removing the except for the authenticate method at the bottom of the
controller. Also changed the name of the local database.

Next step will be to start linking things to the front end, and testing
the front ajax request to make sure they are working correctly.