Add show only user's blogs function
Created backend functionality that will allow a signed in user to
click the index blogs button and view only the blogs that they've
created. Accomplished this by first creating a function in
app/controllers/blogs.js that executes a find method on blogs where
the owner of the blog is the current user that is signed in. After that,
I altered config/routes.js to define route that when called performs
a GET request to /userblogs that execs the previously stated function.
I then created a curl script to test the route and it appears to be
functioning correctly. Next step will be to implement this method on
the front end.